requirements.txt
albucore==0.0.17
albumentations==1.4.16
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
eval_type_backport==0.2.0
filelock==3.16.1
Flask==0.12
Flask-Login==0.4.1
Flask-Mobility==0.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.35.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
lazy_loader==0.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
-e git+https://github.com/IlyaOvodov/OvoTools.git@8a0c8960eed49b19a14b02462886555b084a9a78#egg=ovotools
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.11
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-build==0.18.1
scikit-image==0.24.0
scipy==1.14.1
selenium==4.25.0
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
tifffile==2024.9.20
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==0.14.1
wheel==0.44.0
wsproto==1.2.0
WTForms==2.2.1
